Refactor the folder structure SN-455
